Remove some fixmes (now TODOs in PR) and rename the autocomplete generator as its specific
